rename post hook to auth service, add logout method
